+++ ChangeLog.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [4.11.0] - 2024-03-17
+## [4.12.0] - 2024-03-19
 
 ### Added